+++ internal/cloud/backend_plan_test.go
@@ -49,7 +49,7 @@ func testOperationPlanWithTimeout(t *testing.T, configDir string, timeout time.D
 
 	streams, done := terminal.StreamsForTesting(t)
 	view := views.NewView(streams)
-	stateLockerView := views.NewStateLocker(arguments.ViewHuman, view)
+	stateLockerView := views.NewStateLocker(arguments.ViewOptions{ViewType: arguments.ViewHuman}, view)
 	operationView := views.NewOperation(arguments.ViewHuman, false, view)
 
 	// Many of our tests use an overridden "null" provider that's just in-memory