+++ internal/command/arguments/test_test.go
@@ -53,7 +53,7 @@ func TestParseTest_Vars(t *testing.T) {
 
 	for name, tc := range tcs {
 		t.Run(name, func(t *testing.T) {
-			got, diags := ParseTest(tc.args)
+			got, _, diags := ParseTest(tc.args)
 			if len(diags) > 0 {
 				t.Fatalf("unexpected diags: %v", diags)
 			}
@@ -78,7 +78,7 @@ func TestParseTest(t *testing.T) {
 			want: &Test{
 				Filter:        nil,
 				TestDirectory: "tests",
-				ViewType:      ViewHuman,
+				ViewOptions:   ViewOptions{ViewType: ViewHuman},
 				Vars:          &Vars{},
 			},
 			wantDiags: nil,
@@ -88,7 +88,7 @@ func TestParseTest(t *testing.T) {
 			want: &Test{
 				Filter:        []string{"one.tftest.hcl", "two.tftest.hcl"},
 				TestDirectory: "tests",
-				ViewType:      ViewHuman,
+				ViewOptions:   ViewOptions{ViewType: ViewHuman},
 				Vars:          &Vars{},
 			},
 			wantDiags: nil,
@@ -98,7 +98,7 @@ func TestParseTest(t *testing.T) {
 			want: &Test{
 				Filter:        nil,
 				TestDirectory: "tests",
-				ViewType:      ViewJSON,
+				ViewOptions:   ViewOptions{ViewType: ViewJSON},
 				Vars:          &Vars{},
 			},
 			wantDiags: nil,
@@ -108,7 +108,7 @@ func TestParseTest(t *testing.T) {
 			want: &Test{
 				Filter:        nil,
 				TestDirectory: "other",
-				ViewType:      ViewHuman,
+				ViewOptions:   ViewOptions{ViewType: ViewHuman},
 				Vars:          &Vars{},
 			},
 			wantDiags: nil,
@@ -118,7 +118,7 @@ func TestParseTest(t *testing.T) {
 			want: &Test{
 				Filter:        nil,
 				TestDirectory: "tests",
-				ViewType:      ViewHuman,
+				ViewOptions:   ViewOptions{ViewType: ViewHuman},
 				Verbose:       true,
 				Vars:          &Vars{},
 			},
@@ -128,7 +128,7 @@ func TestParseTest(t *testing.T) {
 			want: &Test{
 				Filter:        nil,
 				TestDirectory: "tests",
-				ViewType:      ViewHuman,
+				ViewOptions:   ViewOptions{ViewType: ViewHuman},
 				Vars:          &Vars{},
 			},
 			wantDiags: tfdiags.Diagnostics{
@@ -141,11 +141,11 @@ func TestParseTest(t *testing.T) {
 		},
 	}
 
-	cmpOpts := cmpopts.IgnoreUnexported(Operation{}, Vars{}, State{})
+	cmpOpts := cmpopts.IgnoreUnexported(Operation{}, Vars{}, State{}, ViewOptions{})
 
 	for name, tc := range tcs {
 		t.Run(name, func(t *testing.T) {
-			got, diags := ParseTest(tc.args)
+			got, _, diags := ParseTest(tc.args)
 
 			if diff := cmp.Diff(tc.want, got, cmpOpts); len(diff) > 0 {
 				t.Errorf("diff:\n%s", diff)