+++ internal/command/output.go
@@ -30,7 +30,8 @@ func (c *OutputCommand) Run(rawArgs []string) int {
 	c.View.Configure(common)
 
 	// Parse and validate flags
-	args, diags := arguments.ParseOutput(rawArgs)
+	args, closer, diags := arguments.ParseOutput(rawArgs)
+	defer closer()
 	if diags.HasErrors() {
 		c.View.Diagnostics(diags)
 		c.View.HelpPrompt("output")
@@ -39,7 +40,7 @@ func (c *OutputCommand) Run(rawArgs []string) int {
 
 	c.View.SetShowSensitive(args.ShowSensitive)
 
-	view := views.NewOutput(args.ViewType, c.View)
+	view := views.NewOutput(args.ViewOptions, c.View)
 
 	// Inject variables from args into meta for static evaluation
 	c.GatherVariables(args.Vars)
@@ -139,33 +140,38 @@ Usage: tofu [global options] output [options] [NAME]
 
 Options:
 
-  -state=path        Path to the state file to read. Defaults to
-                     "terraform.tfstate". Ignored when remote 
-                     state is used.
+  -state=path          Path to the state file to read. Defaults to
+                       "terraform.tfstate". Ignored when remote 
+                       state is used.
+                      
+  -no-color            If specified, output won't contain any color.
+                      
+  -json                If specified, machine readable output will be
+                       printed in JSON format.
 
-  -no-color          If specified, output won't contain any color.
+  -json-into=out.json  Produce the same output as -json, but sent directly
+                       to the given file. This allows automation to preserve
+                       the original human-readable output streams, while
+                       capturing more detailed logs for machine analysis.
 
-  -json              If specified, machine readable output will be
-                     printed in JSON format.
-
-  -raw               For value types that can be automatically
-                     converted to a string, will print the raw
-                     string directly, rather than a human-oriented
-                     representation of the value.
-
-					 Use this with care when stdout is a terminal and when
-					 the output value might contain control characters.
-
-  -show-sensitive    If specified, sensitive values will be displayed.
-
-  -var 'foo=bar'     Set a value for one of the input variables in the root
-                     module of the configuration. Use this option more than
-                     once to set more than one variable.
-
-  -var-file=filename Load variable values from the given file, in addition
-                     to the default files terraform.tfvars and *.auto.tfvars.
-                     Use this option more than once to include more than one
-                     variables file.
+  -raw                 For value types that can be automatically
+                       converted to a string, will print the raw
+                       string directly, rather than a human-oriented
+                       representation of the value.
+                       
+                       Use this with care when stdout is a terminal and when
+                       the output value might contain control characters.
+                       
+  -show-sensitive      If specified, sensitive values will be displayed.
+                       
+  -var 'foo=bar'       Set a value for one of the input variables in the root
+                       module of the configuration. Use this option more than
+                       once to set more than one variable.
+                       
+  -var-file=filename   Load variable values from the given file, in addition
+                       to the default files terraform.tfvars and *.auto.tfvars.
+                       Use this option more than once to include more than one
+                       variables file.
 `
 	return strings.TrimSpace(helpText)
 }