+++ internal/command/plan.go
@@ -37,13 +37,14 @@ func (c *PlanCommand) Run(rawArgs []string) int {
 	c.Meta.Color = c.Meta.color
 
 	// Parse and validate flags
-	args, diags := arguments.ParsePlan(rawArgs)
+	args, closer, diags := arguments.ParsePlan(rawArgs)
+	defer closer()
 
 	c.View.SetShowSensitive(args.ShowSensitive)
 
 	// Instantiate the view, even if there are flag errors, so that we render
 	// diagnostics according to the desired view
-	view := views.NewPlan(args.ViewType, c.View)
+	view := views.NewPlan(args.ViewOptions, c.View)
 
 	if diags.HasErrors() {
 		view.Diagnostics(diags)
@@ -62,7 +63,7 @@ func (c *PlanCommand) Run(rawArgs []string) int {
 	// FIXME: the -input flag value is needed to initialize the backend and the
 	// operation, but there is no clear path to pass this value down, so we
 	// continue to mutate the Meta object state for now.
-	c.Meta.input = args.InputEnabled
+	c.Meta.input = args.ViewOptions.InputEnabled
 
 	// FIXME: the -parallelism flag is used to control the concurrency of
 	// OpenTofu operations. At the moment, this value is used both to
@@ -86,7 +87,7 @@ func (c *PlanCommand) Run(rawArgs []string) int {
 	}
 
 	// Prepare the backend with the backend-specific arguments
-	be, beDiags := c.PrepareBackend(ctx, args.State, args.ViewType, enc)
+	be, beDiags := c.PrepareBackend(ctx, args.State, args.ViewOptions, enc)
 	diags = diags.Append(beDiags)
 	if diags.HasErrors() {
 		view.Diagnostics(diags)
@@ -94,7 +95,7 @@ func (c *PlanCommand) Run(rawArgs []string) int {
 	}
 
 	// Build the operation request
-	opReq, opDiags := c.OperationRequest(ctx, be, view, args.ViewType, args.Operation, args.OutPath, args.GenerateConfigPath, enc)
+	opReq, opDiags := c.OperationRequest(ctx, be, view, args.ViewOptions, args.Operation, args.OutPath, args.GenerateConfigPath, enc)
 	diags = diags.Append(opDiags)
 	if diags.HasErrors() {
 		view.Diagnostics(diags)
@@ -124,7 +125,7 @@ func (c *PlanCommand) Run(rawArgs []string) int {
 	return op.Result.ExitStatus()
 }
 
-func (c *PlanCommand) PrepareBackend(ctx context.Context, args *arguments.State, viewType arguments.ViewType, enc encryption.Encryption) (backend.Enhanced, tfdiags.Diagnostics) {
+func (c *PlanCommand) PrepareBackend(ctx context.Context, args *arguments.State, viewOptions arguments.ViewOptions, enc encryption.Encryption) (backend.Enhanced, tfdiags.Diagnostics) {
 	// FIXME: we need to apply the state arguments to the meta object here
 	// because they are later used when initializing the backend. Carving a
 	// path to pass these arguments to the functions that need them is
@@ -138,8 +139,8 @@ func (c *PlanCommand) PrepareBackend(ctx context.Context, args *arguments.State,
 
 	// Load the backend
 	be, beDiags := c.Backend(ctx, &BackendOpts{
-		Config:   backendConfig,
-		ViewType: viewType,
+		Config:      backendConfig,
+		ViewOptions: viewOptions,
 	}, enc.State())
 	diags = diags.Append(beDiags)
 	if beDiags.HasErrors() {
@@ -153,7 +154,7 @@ func (c *PlanCommand) OperationRequest(
 	ctx context.Context,
 	be backend.Enhanced,
 	view views.Plan,
-	viewType arguments.ViewType,
+	viewOptions arguments.ViewOptions,
 	args *arguments.Operation,
 	planOutPath string,
 	generateConfigOut string,
@@ -162,7 +163,7 @@ func (c *PlanCommand) OperationRequest(
 	var diags tfdiags.Diagnostics
 
 	// Build the operation
-	opReq := c.Operation(ctx, be, viewType, enc)
+	opReq := c.Operation(ctx, be, viewOptions, enc)
 	opReq.ConfigDir = "."
 	opReq.PlanMode = args.PlanMode
 	opReq.Hooks = view.Hooks()
@@ -328,6 +329,11 @@ Other Options:
   -json                        Produce output in a machine-readable JSON
                                format, suitable for use in text editor
                                integrations and other automated systems.
+
+  -json-into=out.json          Produce the same output as -json, but sent directly
+                               to the given file. This allows automation to preserve
+                               the original human-readable output streams, while
+                               capturing more detailed logs for machine analysis.
 
   -deprecation=module:m        Specify what type of warnings are shown.
                                Accepted values for "m": all, local, none.