+++ internal/command/state_push.go
@@ -117,7 +117,7 @@ func (c *StatePushCommand) Run(args []string) int {
 	}
 
 	if c.stateLock {
-		stateLocker := clistate.NewLocker(c.stateLockTimeout, views.NewStateLocker(arguments.ViewHuman, c.View))
+		stateLocker := clistate.NewLocker(c.stateLockTimeout, views.NewStateLocker(arguments.ViewOptions{ViewType: arguments.ViewHuman}, c.View))
 		if diags := stateLocker.Lock(stateMgr, "state-push"); diags.HasErrors() {
 			c.showDiagnostics(diags)
 			return 1